Store column numbers as a u32
It's reasonable to assume that a single column won't exceed four
billion, and this is a small speedup (1% instruction count reduction
on the sample file in #293).

+++ src/lines.rs
@@ -102,12 +102,12 @@ impl NewlinePositions {
                     0
                 } else {
                     region_start - line_start
-                },
+                } as u32,
                 end_col: if region_end < line_end {
                     region_end - line_start
                 } else {
                     line_end - line_start
-                },
+                } as u32,
             });
         }
